식단조회시 리뷰평점수정

+++ domain-makers/src/main/java/co/dalicious/domain/food/dto/DailyFoodDto.java
@@ -25,7 +25,7 @@ public class DailyFoodDto {
     private String spicy;
     private String vegan;
     private String image;
-    private Integer reviewAverage;
+    private double reviewAverage;
     private String description;
     private BigDecimal price;
     private BigDecimal discountedPrice;
